# api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: LibroRed API
  description: REST API for LibroRed - Web Application for Book Lending Between Individuals
  contact:
    name: Team 13
    url: https://github.com/medinaymedia
    email: am.juradoc@alumnos.urjc.es
  license:
    name: Apache 2.0
    url: http://springdoc.org
  version: '1.0'
servers:
- url: https://localhost:8443
  description: Development server
...
paths:
...
  /api/v1/books:
    get:
      tags:
      - Books
      summary: Get all books
      description: Retrieve a list of all books
      operationId: getAllBooks
      responses:
        '200':
          description: Books retrieved successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookDTO'
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookDTO'
    post:
      tags:
      - Books
      summary: Create a new book
      description: Create a new book entry
      operationId: createBook
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookDTO'
        required: true
      responses:
        '201':
          description: Book created successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BookDTO'
        '400':
          description: Invalid book data
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BookDTO'
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BookDTO'
...
  /api/books:
    get:
      tags:
      - Books
      summary: Get books with pagination
      description: Retrieve paginated list of books with optional filtering
      operationId: getBooksPage
      parameters:
      - name: page
        in: query
        description: Page number (0-based)
        required: false
        schema:
          type: integer
          default: 0
      - name: size
        in: query
        description: Page size
        required: false
        schema:
          type: integer
          default: 6
      - name: genre
        in: query
        description: Filter by genre
        required: false
        schema:
          type: string
      - name: search
        in: query
        description: Search term for title or author
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Books retrieved successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BookPageResponse'
    post:
      tags:
      - Books
      summary: Create a new book
      description: Create a new book entry (authenticated users only)
      operationId: createBookPost
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BookCreateRequest'
        required: true
      responses:
        '201':
          description: Book created successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BookDTO'
        '400':
          description: Invalid book data
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BookDTO'
        '401':
          description: User not authenticated
  /api/books/{id}/cover:
    get:
      tags:
      - Books
      summary: Get book cover image
      description: Retrieve book cover image by book ID
      operationId: getBookCover
      parameters:
      - name: id
        in: path
        description: Book ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Book cover image
          content:
            image/*:
              schema:
                type: string
                format: binary
        '404':
          description: Book or cover not found
...
components:
  schemas:
...
    BookDTO:
      type: object
      description: Book data
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        author:
          type: string
        genre:
          type: string
          enum:
          - Fiction
          - Non_Fiction
          - Mystery_Thriller
          - SciFi_Fantasy
          - Romance
          - Historical_Fiction
          - Horror
        description:
          type: string
        hasCoverImage:
          type: boolean
        owner:
          $ref: '#/components/schemas/UserBasicDTO'
...
    UserBasicDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
...
    BookPageResponse:
      type: object
      description: Paginated book response
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/BookDTO'
        totalPages:
          type: integer
        totalElements:
          type: integer
        size:
          type: integer
        number:
          type: integer
        hasNext:
          type: boolean
    BookCreateRequest:
      type: object
      description: Book creation request
      properties:
        title:
          type: string
        author:
          type: string
        genre:
          type: string
          enum:
          - Fiction
          - Non_Fiction
          - Mystery_Thriller
          - SciFi_Fantasy
          - Romance
          - Historical_Fiction
          - Horror
        description:
          type: string
        coverImage:
          type: string
          format: binary
      required:
        - title
        - author
        - genre